requirements.txt
pandas>=2.1.0
SQLAlchemy>=2.0.0
oracledb>=2.0.0
PyYAML>=6.0.1
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.13.0
fpdf2>=2.7.9
schedule>=1.2.0
streamlit>=1.34.0
pytest>=7.4.0
